refactor: KichenridersClient를 model -> repsoitory 이동
